Move dm_group_message() to KoalaBot.py and update tests in test_KoalaBot.py
